refactor: move ListAsMap
